@@ -1,14 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion '26.0.2'
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.telink.bluetooth.light"
-        minSdkVersion 18
+        minSdkVersion 19
         targetSdkVersion 26
-        versionCode 6
-        versionName "2.5"
+        versionCode 7
+        versionName "2.6.0"
     }
     buildTypes {
         release {
@@ -31,12 +30,12 @@ repositories {
 }
 
 dependencies {
-    //compile(name: "libTelinkLight", ext: "aar")
-//    compile 'com.android.support:support-v4:26.1.0'
-    compile 'com.android.support:recyclerview-v7:26.1.0'
-    compile project(':BluetoothLightLib')
-    compile 'com.google.zxing:core:3.2.1'
-    compile 'com.google.zxing:android-core:3.2.1'
-    //compile project(':BluetoothStorageLib')
-    compile 'com.google.code.gson:gson:2.8.2'
+//    implementation 'com.android.support:recyclerview-v7:26.1.0'
+    implementation project(':BluetoothLightLib')
+    implementation 'cn.bingoogolapple:bga-qrcode-zxing:1.3.6'
+    implementation 'com.google.zxing:core:3.2.1'
+    implementation 'com.google.zxing:android-core:3.2.1'
+    implementation 'com.google.code.gson:gson:2.8.2'
+//    implementation 'com.android.support:design:26.1.0'
+    implementation 'com.google.android.material:material:1.0.0'
 }
