@@ -1,13 +1,13 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion '26.0.2'
+    compileSdkVersion 29
+
     defaultConfig {
-        minSdkVersion 18
+        minSdkVersion 19
         targetSdkVersion 26
         versionCode 6
-        versionName "2.5"
+        versionName "2.6.0"
     }
     buildTypes {
         release {
@@ -24,5 +24,7 @@ android {
 }
 
 dependencies {
-    compile 'com.android.support:support-v13:26.1.0'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.localbroadcastmanager:localbroadcastmanager:1.0.0'
+//    compile 'com.android.support:support-v13:26.1.0'
 }
