@@ -1,14 +1,14 @@
 /********************************************************************************************************
- * @file     TelinkBaseActivity.java 
+ * @file TelinkBaseActivity.java
  *
- * @brief    for TLSR chips
+ * @brief for TLSR chips
  *
- * @author	 telink
- * @date     Sep. 30, 2010
+ * @author telink
+ * @date Sep. 30, 2010
  *
- * @par      Copyright (c) 2010, Telink Semiconductor (Shanghai) Co., Ltd.
+ * @par Copyright (c) 2010, Telink Semiconductor (Shanghai) Co., Ltd.
  *           All rights reserved.
- *           
+ *
  *			 The information contained herein is confidential and proprietary property of Telink 
  * 		     Semiconductor (Shanghai) Co., Ltd. and is available under the terms 
  *			 of Commercial License Agreement between Telink Semiconductor (Shanghai) 
@@ -17,16 +17,20 @@
  *
  * 			 Licensees are granted free, non-transferable use of the information in this 
  *			 file under Mutual Non-Disclosure Agreement. NO WARRENTY of ANY KIND is provided. 
- *           
+ *
  *******************************************************************************************************/
 package com.telink.bluetooth.light;
 
 import android.annotation.SuppressLint;
-import android.app.Activity;
 import android.os.Bundle;
+import android.view.View;
+import android.widget.TextView;
 import android.widget.Toast;
 
-public class TelinkBaseActivity extends Activity {
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+
+public class TelinkBaseActivity extends AppCompatActivity {
 
     protected Toast toast;
     protected boolean foreground = false;
@@ -68,6 +72,27 @@ public class TelinkBaseActivity extends Activity {
             this.toast.setText(s);
             this.toast.show();
         }
+    }
+
+    protected void setTitle(String title) {
+        TextView tv_title = findViewById(R.id.tv_title);
+        if (tv_title != null) {
+            tv_title.setText(title);
+        }
+    }
+
+    protected void enableBackNav(boolean enable) {
+        Toolbar toolbar = findViewById(R.id.title_bar);
+        if (enable) {
+            toolbar.setNavigationOnClickListener(new View.OnClickListener() {
+                public void onClick(View v) {
+                    finish();
+                }
+            });
+        } else {
+            toolbar.setNavigationIcon(null);
+        }
+
     }
 
     protected void saveLog(String log) {
