@@ -22,7 +22,7 @@
 package com.telink.bluetooth.light.activity;
 
 import android.Manifest;
-import android.app.AlertDialog;
+import androidx.appcompat.app.AlertDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageManager;
@@ -31,12 +31,13 @@ import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
 import android.provider.Settings;
-import android.support.annotation.NonNull;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
 
 import com.telink.bluetooth.light.R;
 import com.telink.bluetooth.light.TelinkBaseActivity;
+
+import androidx.annotation.NonNull;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 
 
 /**
@@ -46,7 +47,8 @@ import com.telink.bluetooth.light.TelinkBaseActivity;
 
 public class SplashActivity extends TelinkBaseActivity {
 
-    private static final int PERMISSIONS_REQUEST_LOCATION = 0x10;
+    //    private static final int PERMISSIONS_REQUEST_LOCATION = 0x10;
+    private static final int PERMISSIONS_REQUEST_ALL = 0x10;
 
     private Handler delayHandler = new Handler();
 
@@ -62,17 +64,23 @@ public class SplashActivity extends TelinkBaseActivity {
     @Override
     protected void onStart() {
         super.onStart();
-        if (settingDialog != null){
+        if (settingDialog != null) {
             settingDialog.dismiss();
         }
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            if (ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION)
-                    == PackageManager.PERMISSION_GRANTED) {
+            if (ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED
+                    &&
+                    ContextCompat.checkSelfPermission(this, Manifest.permission.READ_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED
+                    &&
+                    ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
                 onPermissionChecked();
             } else {
                 ActivityCompat.requestPermissions(this,
-                        new String[]{Manifest.permission.ACCESS_FINE_LOCATION},
-                        PERMISSIONS_REQUEST_LOCATION);
+                        new String[]{
+                                Manifest.permission.ACCESS_FINE_LOCATION,
+                                Manifest.permission.READ_EXTERNAL_STORAGE,
+                                Manifest.permission.WRITE_EXTERNAL_STORAGE},
+                        PERMISSIONS_REQUEST_ALL);
             }
         } else {
             onPermissionChecked();
@@ -95,7 +103,7 @@ public class SplashActivity extends TelinkBaseActivity {
         if (settingDialog == null) {
             AlertDialog.Builder dialogBuilder = new AlertDialog.Builder(this);
             dialogBuilder.setCancelable(false);
-            dialogBuilder.setTitle("Warn");
+            dialogBuilder.setTitle("Warning");
             dialogBuilder.setMessage("Location permission is necessary when searching bluetooth device on 6.0 or upper device");
             dialogBuilder.setPositiveButton("Go Settings", new DialogInterface.OnClickListener() {
                 @Override
@@ -127,12 +135,12 @@ public class SplashActivity extends TelinkBaseActivity {
     @Override
     public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        if (requestCode == PERMISSIONS_REQUEST_LOCATION) {
+        if (requestCode == PERMISSIONS_REQUEST_ALL) {
             if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
                 onPermissionChecked();
             } else {
                 onPermissionDenied();
-                showToast("Location Permission Denied");
+                showToast("Permission Denied");
             }
         }
 
