@@ -22,7 +22,6 @@
 package com.telink.bluetooth.light.adapter;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -34,6 +33,8 @@ import com.telink.bluetooth.light.R;
 import com.telink.bluetooth.light.model.Light;
 
 import java.util.List;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * Created by kee on 2017/12/19.
