@@ -1,25 +1,5 @@
-/********************************************************************************************************
- * @file     FileListAdapter.java 
- *
- * @brief    for TLSR chips
- *
- * @author	 telink
- * @date     Sep. 30, 2010
- *
- * @par      Copyright (c) 2010, Telink Semiconductor (Shanghai) Co., Ltd.
- *           All rights reserved.
- *           
- *			 The information contained herein is confidential and proprietary property of Telink 
- * 		     Semiconductor (Shanghai) Co., Ltd. and is available under the terms 
- *			 of Commercial License Agreement between Telink Semiconductor (Shanghai) 
- *			 Co., Ltd. and the licensee in separate contract or the terms described here-in. 
- *           This heading MUST NOT be removed from this file.
- *
- * 			 Licensees are granted free, non-transferable use of the information in this 
- *			 file under Mutual Non-Disclosure Agreement. NO WARRENTY of ANY KIND is provided. 
- *           
- *******************************************************************************************************/
-package com.telink.bluetooth.light.activity;
+
+package com.telink.bluetooth.light.file;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -37,21 +17,28 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Created by Administrator on 2016/10/8.
+ * Created by ke on 2016/10/8.
  */
 public class FileListAdapter extends BaseAdapter {
 
     private Context mContext;
     private List<File> mFiles;
+    private String targetSuffix;
+    private boolean searchMode = false;
 
-    FileListAdapter(Context context) {
+    FileListAdapter(Context context, String targetSuffix) {
         this.mContext = context;
+        this.targetSuffix = targetSuffix;
         mFiles = new ArrayList<>();
     }
 
-    public void setData(List<File> files){
+    public void setData(List<File> files) {
         this.mFiles = files;
         this.notifyDataSetChanged();
+    }
+
+    public void setSearchMode(boolean searchMode) {
+        this.searchMode = searchMode;
     }
 
     @Override
@@ -73,9 +60,10 @@ public class FileListAdapter extends BaseAdapter {
     public View getView(int position, View convertView, ViewGroup parent) {
         ViewHolder holder;
         if (convertView == null) {
-            convertView = LayoutInflater.from(mContext).inflate(R.layout.item_file_list, null);
+            convertView = LayoutInflater.from(mContext).inflate(R.layout.item_file, null);
             holder = new ViewHolder();
             holder.tv_name = (TextView) convertView.findViewById(R.id.tv_name);
+            holder.tv_path = (TextView) convertView.findViewById(R.id.tv_path);
             holder.iv_icon = (ImageView) convertView.findViewById(R.id.iv_icon);
             holder.iv_right = (ImageView) convertView.findViewById(R.id.iv_right);
             convertView.setTag(holder);
@@ -83,19 +71,25 @@ public class FileListAdapter extends BaseAdapter {
             holder = (ViewHolder) convertView.getTag();
         }
         if (mFiles.get(position).isDirectory()) {
-            // 目录
-            holder.iv_icon.setImageResource(R.drawable.folder);
+            // dir
+            holder.iv_icon.setImageResource(R.drawable.ic_folder);
             holder.iv_right.setVisibility(View.VISIBLE);
         } else {
             holder.iv_right.setVisibility(View.GONE);
-            // 文件
-            if (mFiles.get(position).getName().endsWith(".bin")) {
-                holder.iv_icon.setImageResource(R.drawable.file_right);
+            // file .bin
+            if (targetSuffix != null && !targetSuffix.equals("") && mFiles.get(position).getName().endsWith(targetSuffix)) {
+                holder.iv_icon.setImageResource(R.drawable.ic_file_checked);
             } else {
-                holder.iv_icon.setImageResource(R.drawable.file);
+                holder.iv_icon.setImageResource(R.drawable.ic_file_unchecked);
             }
         }
         holder.tv_name.setText(mFiles.get(position).getName());
+        if (searchMode) {
+            holder.tv_path.setVisibility(View.VISIBLE);
+            holder.tv_path.setText(mFiles.get(position).getAbsolutePath());
+        } else {
+            holder.tv_path.setVisibility(View.GONE);
+        }
         return convertView;
     }
 
@@ -103,5 +97,6 @@ public class FileListAdapter extends BaseAdapter {
         public TextView tv_name;
         public ImageView iv_icon;
         public ImageView iv_right;
+        public TextView tv_path;
     }
 }
