@@ -1,14 +1,14 @@
 /********************************************************************************************************
- * @file     Lights.java 
+ * @file Lights.java
  *
- * @brief    for TLSR chips
+ * @brief for TLSR chips
  *
- * @author	 telink
- * @date     Sep. 30, 2010
+ * @author telink
+ * @date Sep. 30, 2010
  *
- * @par      Copyright (c) 2010, Telink Semiconductor (Shanghai) Co., Ltd.
+ * @par Copyright (c) 2010, Telink Semiconductor (Shanghai) Co., Ltd.
  *           All rights reserved.
- *           
+ *
  *			 The information contained herein is confidential and proprietary property of Telink 
  * 		     Semiconductor (Shanghai) Co., Ltd. and is available under the terms 
  *			 of Commercial License Agreement between Telink Semiconductor (Shanghai) 
@@ -17,7 +17,7 @@
  *
  * 			 Licensees are granted free, non-transferable use of the information in this 
  *			 file under Mutual Non-Disclosure Agreement. NO WARRENTY of ANY KIND is provided. 
- *           
+ *
  *******************************************************************************************************/
 package com.telink.bluetooth.light.model;
 
@@ -26,7 +26,6 @@ import android.text.TextUtils;
 import com.telink.bluetooth.light.ConnectionStatus;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
@@ -70,7 +69,7 @@ public class Lights extends DataStorageImpl<Light> {
         }
         List<Light> localList = new ArrayList<>();
         for (Light light : data) {
-            if (!TextUtils.isEmpty(light.macAddress) && (online && light.connectionStatus != ConnectionStatus.OFFLINE)) {
+            if (!TextUtils.isEmpty(light.macAddress) && (!online || light.connectionStatus != ConnectionStatus.OFFLINE)) {
                 localList.add(light);
             }
         }
