@@ -1,14 +1,14 @@
 /********************************************************************************************************
- * @file     AdvanceStrategy.java 
+ * @file AdvanceStrategy.java
  *
- * @brief    for TLSR chips
+ * @brief for TLSR chips
  *
- * @author	 telink
- * @date     Sep. 30, 2010
+ * @author telink
+ * @date Sep. 30, 2010
  *
- * @par      Copyright (c) 2010, Telink Semiconductor (Shanghai) Co., Ltd.
+ * @par Copyright (c) 2010, Telink Semiconductor (Shanghai) Co., Ltd.
  *           All rights reserved.
- *           
+ *
  *			 The information contained herein is confidential and proprietary property of Telink 
  * 		     Semiconductor (Shanghai) Co., Ltd. and is available under the terms 
  *			 of Commercial License Agreement between Telink Semiconductor (Shanghai) 
@@ -17,7 +17,7 @@
  *
  * 			 Licensees are granted free, non-transferable use of the information in this 
  *			 file under Mutual Non-Disclosure Agreement. NO WARRENTY of ANY KIND is provided. 
- *           
+ *
  *******************************************************************************************************/
 package com.telink.bluetooth.light;
 
@@ -217,8 +217,11 @@ public abstract class AdvanceStrategy {
             }
 
             if (now && this.mCallback != null) {
-                Log.d(TAG, "Sample Opcode : " + Integer.toHexString(opcode & 0xFF) + " delay:" + delay);
-
+                if (address == 0 && opcode == (byte)0xDC) {
+                    // direct command
+                    Log.d(TAG, String.format("Sample - direct command  opcode : %02X delay : " + delay, opcode));
+                    return this.mCallback.onCommandSampled(opcode, address, params, tag, delay);
+                }
                 long during = currentTime - this.lastCmdTime;
                 if (during < 0) {
                     delay = COMMAND_DELAY;
@@ -231,6 +234,7 @@ public abstract class AdvanceStrategy {
                 } else {
                     this.lastCmdTime = currentTime + delay;
                 }
+                Log.d(TAG, "Sample Opcode : " + Integer.toHexString(opcode & 0xFF) + " delay:" + delay);
 /*
                 long period = currentTime - this.lastCmdTime;
                 if (period > 0 && period < COMMAND_DELAY) {
